refactor : 클래스 기능별로 패키지 분리
