+++ src/main/java/calculator/input/InputController.java
@@ -1,4 +1,6 @@
-package calculator;
+package calculator.input;
+
+import calculator.data.InputString;
 
 import static camp.nextstep.edu.missionutils.Console.close;
 import static camp.nextstep.edu.missionutils.Console.readLine;
