+++ src/main/java/calculator/output/PrintOutput.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.output;
 
 public class PrintOutput {
