+++ src/main/java/calculator/service/InputParser.java
@@ -1,4 +1,6 @@
-package calculator;
+package calculator.service;
+
+import calculator.validation.CheckForm;
 
 import java.util.ArrayList;
 import java.util.List;
