+++ src/main/java/calculator/service/Parser.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.service;
 
 import java.util.List;
 import java.util.Set;
