+++ src/main/java/calculator/service/StringPlusCalculator.java
@@ -1,4 +1,6 @@
-package calculator;
+package calculator.service;
+
+import calculator.data.InputString;
 
 public class StringPlusCalculator {
