+++ src/main/java/calculator/validation/CheckForm.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.validation;
 
 import java.util.Set;
